# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Customer {
  id: string;
  name: string;
  phone: string;
  createdAt: string;
}

export interface CreateCustomerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
}

export interface MessageResponse {
  message: string;
}

export interface UpdateCustomerInput {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phone?: string;
}

export interface ErrorResponse {
  error: string;
}