# lib/api-zod/src/generated/types/updateCustomerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCustomerInput {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  phone?: string;
}